pyproject.toml
[project]
name = "newsletterchat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.3.24",
    "aiosqlite>=0.22.1",
    "alembic>=1.18.1",
    "brave-search>=0.1.0",
    "dependency-injector>=4.48.3",
    "fastapi>=0.128.0",
    "groq>=1.0.0",
    "passlib[bcrypt]>=1.7.4",
    "pika>=1.3.2",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.45",
    "uvicorn[standard]>=0.40.0",
]
# uv run pytest --cov=app --cov-report=html
# uv add --group dev factory-boy
[dependency-groups]
dev = [
    "factory-boy>=3.3.0",
    "httpx>=0.25.2",
    "pytest>=8.4.2",
    "pytest-asyncio>=0.23.8",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.11",
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"

[tool.ruff]
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
    "htmlcov",
    "logs",
    ".pytest_cache",
    "alembic",
]

# Mesmo que Black
line-length = 88
indent-width = 4

# Versão alvo do Python
target-version = "py312"

[tool.ruff.lint]
# Habilitar regras padrão: pycodestyle (E), Pyflakes (F), etc.
select = ["E4", "E7", "E9", "F", "B", "I", "UP", "SIM"]

# Ignorar regras específicas, como comprimento de linha
ignore = ["E501"]

# Permitir correção automática para todas as regras habilitadas
fixable = ["ALL"]

# Regras que não podem ser corrigidas automaticamente
unfixable = ["B"]

[tool.ruff.format]
# Usar aspas duplas para strings, como Black
quote-style = "double"

# Indentar com espaços
indent-style = "space"

# Detectar automaticamente o final de linha
line-ending = "auto"

# Formatar código em docstrings
docstring-code-format = true

# Como Usar o Ruff:
# Verificar o código: Execute uv run ruff check . para identificar problemas de linting.
# Corrigir automaticamente: Use uv run ruff check --fix . para corrigir problemas fixáveis (como organização de imports).
# Formatar o código: Execute uv run ruff format . para formatar o código de acordo com as regras definidas.
